requirements.txt
Django<5
Pillow
psycopg2
djangorestframework
djangorestframework-simplejwt
django_ratelimit